backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-socketio==5.11.0
apscheduler==3.10.4
python-dotenv==1.0.1
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-openai==0.2.7
langchain-core==0.3.21
pydantic==2.9.0
pydantic-settings==2.5.0
